put mapDataManipulation consts in their own file
