+++ src/mapDataManipulation/const.ts
@@ -1,0 +1,10 @@
+export type PredefinedEffects = {
+  gain?: number;
+  gamma?: number;
+};
+
+export type RgbMappingArrays = {
+  red: number[];
+  green: number[];
+  blue: number[];
+};
